Initial TypeScript conversion
