+++ config/11ty/collections/breadcrumbs.ts
@@ -1,5 +1,5 @@
 import _ from 'lodash'
-import utils from '../../utils/index.mjs'
+import utils from '@utils/index'
 
 export default (collectionApi) => {
   const cats = utils.categories.tree()
