--- config/11ty/collections/index.mjs
@@ -1,11 +1,0 @@
-import snippetsGrouped from './snippetsGrouped.mjs'
-import breadcrumbs from './breadcrumbs.mjs'
-import relatedSnippets from './related-snippets.mjs'
-import groupData from './groupData.mjs'
-
-export default {
-  relatedSnippets,
-  snippetsGrouped,
-  breadcrumbs,
-  groupData,
-}
