+++ config/11ty/collections/index.ts
@@ -1,0 +1,11 @@
+import snippetsGrouped from './snippetsGrouped'
+import breadcrumbs from './breadcrumbs'
+import relatedSnippets from './related-snippets'
+import groupData from './groupData'
+
+export default {
+  relatedSnippets,
+  snippetsGrouped,
+  breadcrumbs,
+  groupData,
+}
