+++ config/11ty/filters/catPath.ts
@@ -1,3 +1,3 @@
-import utils from '../../utils/index.mjs'
+import utils from '@utils/index'
 
 export default (page) => utils.categories.normalPath({ page })
