--- config/11ty/filters/index.mjs
@@ -1,15 +1,0 @@
-import debugFilter from './debug.mjs'
-import urlize from './urlize.mjs'
-import titlecase from './titlecase.mjs'
-import head from './head.mjs'
-import catPath from './catPath.mjs'
-import markdown from './markdown.mjs'
-
-export default {
-  urlize,
-  titlecase,
-  head,
-  debugFilter,
-  catPath,
-  markdown,
-}
