+++ config/11ty/filters/index.ts
@@ -1,0 +1,15 @@
+import debugFilter from './debug'
+import urlize from './urlize'
+import titlecase from './titlecase'
+import head from './head'
+import catPath from './catPath'
+import markdown from './markdown'
+
+export default {
+  urlize,
+  titlecase,
+  head,
+  debugFilter,
+  catPath,
+  markdown,
+}
