+++ config/11ty/plugins/index.ts
@@ -1,7 +1,7 @@
 import syntaxHighlight from '@11ty/eleventy-plugin-syntaxhighlight'
 import htmlmin from 'html-minifier'
 import { EleventyHtmlBasePlugin } from '@11ty/eleventy'
-import md from './md.mjs'
+import md from './md'
 
 export default {
   syntaxHighlight,
