+++ config/11ty/shortcodes/htmlList.ts
@@ -1,5 +1,5 @@
 import _ from 'lodash'
-import utils from '../../utils/index.mjs'
+import utils from '@utils/index'
 
 const href = (name, link) => (`<a href="${link}">${name}</a>`)
 const path = (base, next) => (!next) ? `${base}/` : `${base}${next}/`
