--- config/11ty/shortcodes/index.mjs
@@ -1,7 +1,0 @@
-import htmlList from './htmlList.mjs'
-import snippetLinks from './snippetLinks.mjs'
-
-export default {
-  htmlList,
-  snippetLinks,
-}
