+++ config/11ty/shortcodes/index.ts
@@ -1,0 +1,7 @@
+import htmlList from './htmlList'
+import snippetLinks from './snippetLinks'
+
+export default {
+  htmlList,
+  snippetLinks,
+}
