+++ config/11ty/templates/categories.ts
@@ -1,5 +1,5 @@
 import _ from 'lodash'
-import utils from '../../utils/index.mjs'
+import utils from '@utils/index'
 
 /**
  * https://www.11ty.dev/docs/permalinks/#use-template-syntax-in-permalink
