+++ config/11ty/templates/groups.ts
@@ -1,5 +1,5 @@
-import utils from '../../utils/index.mjs'
 import _ from 'lodash'
+import utils from '@utils/index'
 
 /**
  * Data Structure
