--- config/11ty/templates/index.mjs
@@ -1,7 +1,0 @@
-import virtualTemplates from './categories.mjs'
-import groups from './groups.mjs'
-
-export default {
-  virtualTemplates,
-  groups,
-}
