+++ config/11ty/templates/index.ts
@@ -1,0 +1,7 @@
+import virtualTemplates from '@templates/categories'
+import groups from '@templates/groups'
+
+export default {
+  virtualTemplates,
+  groups,
+}
