--- config/11ty/transforms/index.mjs
@@ -1,5 +1,0 @@
-import minify from './minify.mjs'
-
-export default {
-  minify,
-}
