+++ config/11ty/transforms/index.ts
@@ -1,0 +1,5 @@
+import minify from './minify'
+
+export default {
+  minify,
+}
